refactor: JdbcTemplate 으로 통합하여 중복 제거
